+++ src/snowcli/cli/constants.py
@@ -6,5 +6,14 @@ DEPLOYMENT_STAGE = "deployments"
 
 
 class ObjectType(Enum):
+    COMPUTE_POOL = "compute pool"
+    DATABASE = "database"
     FUNCTION = "function"
+    JOB = "job"
     PROCEDURE = "procedure"
+    ROLE = "role"
+    SCHEMA = "schema"
+    SERVICE = "service"
+    STREAMLIT = "streamlit"
+    TABLE = "table"
+    WAREHOUSE = "warehouse"
