+++ src/snowcli/cli/object/utils.py
@@ -1,21 +1,6 @@
-from enum import Enum
 from typing import Dict
 
-
-class ObjectType(str, Enum):
-    COMPUTE_POOL = "compute pool"
-    DATABASE = "database"
-    FUNCTION = "function"
-    JOB = "job"
-    PROCEDURE = "procedure"
-    ROLE = "role"
-    SCHEMA = "schema"
-    SERVICE = "service"
-    STREAMLIT = "streamlit"
-    TABLE = "table"
-    WAREHOUSE = "warehouse"
-
-
+from snowcli.cli.constants import ObjectType
 def get_plural_name(object_type: ObjectType):
     exceptions: Dict[str, str] = {}
